+++ test/demoApp/presentation/island.js
@@ -1,0 +1,6 @@
+/**
+ * We use this to test the detection of Orphan nodes
+ */
+export function somethingNotUsed () {
+  // This is an orphan function
+}